Rename and extract handleFormSubmit -> handlePageRequest
